refactored something from yesterday.
