+++ backend/src/main/java/com/htecgroup/skynest/lambda/LambdaType.java
@@ -1,7 +1,8 @@
 package com.htecgroup.skynest.lambda;
 
 public enum LambdaType {
-  UPLOAD_FILE_TO_EXTERNAL_SERVICE_LAMBDA("UFTDL");
+  // The code is used to represent the lambdas in the database and as such should not be changed.
+  UPLOAD_FILE_TO_EXTERNAL_SERVICE_LAMBDA("U_F_T_E_S_L");
 
   private String code;
